projeto_exercicio_flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
werkzeug==2.3.7